# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request OTP
      description: Send a one-time password to the given email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequestOtpBody"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthRequestOtpResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP
      description: Verify the OTP and receive a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthVerifyOtpBody"
      responses:
        "200":
          description: Verification successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthVerifyOtpResponse"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      description: Returns the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /settings/calorie-daily-goal:
    get:
      operationId: getCalorieDailyGoal
      tags: [settings]
      summary: Get calorie daily goal
      responses:
        "200":
          description: Calorie daily goal or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NullableIntGoal"
    post:
      operationId: setCalorieDailyGoal
      tags: [settings]
      summary: Set calorie daily goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntGoalBody"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntGoal"
  /settings/calorie-burn-goal:
    get:
      operationId: getCalorieBurnGoal
      tags: [settings]
      summary: Get calorie burn goal
      responses:
        "200":
          description: Calorie burn goal or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NullableIntGoal"
    post:
      operationId: setCalorieBurnGoal
      tags: [settings]
      summary: Set calorie burn goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntGoalBody"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntGoal"
  /settings/widget-visibility:
    get:
      operationId: getWidgetVisibility
      tags: [settings]
      summary: Get widget visibility settings
      responses:
        "200":
          description: Widget visibility map or null
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/WidgetVisibilityMap"
                  - type: "null"
    put:
      operationId: setWidgetVisibility
      tags: [settings]
      summary: Save widget visibility settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetVisibilityMap"
      responses:
        "200":
          description: Saved widget visibility
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetVisibilityMap"
...
  /settings/insights-date-range:
    get:
      operationId: getInsightsDateRange
      tags: [settings]
      summary: Get Insights date range preference
      description: Returns the saved date range preset for the Insights tab, or null if not set
      responses:
        "200":
          description: Saved date range or null
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/InsightsDateRange"
                  - type: "null"
    put:
      operationId: setInsightsDateRange
      tags: [settings]
      summary: Save Insights date range preference
      description: Persists the selected date range preset for the Insights tab
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsightsDateRange"
      responses:
        "200":
          description: Saved date range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightsDateRange"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /settings/weekly-sessions-goal:
    get:
      operationId: getWeeklySessionsGoal
      tags: [settings]
      summary: Get weekly sessions goal
      description: Returns the target number of workout sessions per week, or null if not set
      responses:
        "200":
          description: Weekly sessions goal or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NullableIntGoal"
    post:
      operationId: setWeeklySessionsGoal
      tags: [settings]
      summary: Set weekly sessions goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntGoalBody"
      responses:
        "200":
          description: Saved goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntGoal"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    NullableIntGoal:
      type: object
      properties:
        value:
          type: ["integer", "null"]
      required:
        - value
    IntGoal:
      type: object
      properties:
        value:
          type: integer
      required:
        - value
    IntGoalBody:
      type: object
      properties:
        value:
          type: integer
      required:
        - value
    WidgetVisibilityMap:
      type: object
      additionalProperties:
        type: boolean
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    InsightsDateRange:
      type: object
      properties:
        preset:
          type: ["string", "null"]
          enum: ["4w", "3m", "6m", "1y", "custom", null]
          description: Date range preset identifier, or null for "All time"
        customStart:
          type: string
          description: Custom range start date (YYYY-MM-DD), used when preset is "custom"
        customEnd:
          type: string
          description: Custom range end date (YYYY-MM-DD), used when preset is "custom"
      required:
        - preset
        - customStart
        - customEnd
...
    AuthRequestOtpBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    AuthVerifyOtpBody:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
        guestToken:
          type: string
      required:
        - email
        - code
    AuthRequestOtpResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    AuthVerifyOtpResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - token
        - user
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: ["string", "null"]
        isGuest:
          type: boolean
      required:
        - id
        - email
        - isGuest
        - customEnd