# lib/api-zod/src/generated/types/insightsDateRange.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InsightsDateRangePreset } from './insightsDateRangePreset';

export interface InsightsDateRange {
  /**
     * Date range preset identifier, or null for "All time"
     * @nullable
     */
  preset: InsightsDateRangePreset;
  /** Custom range start date (YYYY-MM-DD), used when preset is "custom" */
  customStart: string;
  /** Custom range end date (YYYY-MM-DD), used when preset is "custom" */
  customEnd: string;
}